@@ -14,23 +14,30 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
-        "zendframework/zend-loader": "2.*",
-        "zendframework/zend-mvc": "2.*",
-        "zendframework/zend-i18n": "2.*",
-        "zendframework/zend-modulemanager": "2.*",
-        "zendframework/zend-view": "2.*",
-        "zendframework/zend-serializer": "2.*",
-        "zendframework/zend-log": "2.*",
-        "zendframework/zend-servicemanager": "2.*",
-        "zendframework/zend-escaper" : "2.*",
-        "casasoft/casamodules": "dev-master"
+        "php": "^7.4 || ~8.0.0 || ~8.1.0 || ~8.2.0 || ~8.3.0",
+        "laminas/laminas-loader": "^2.0",
+        "laminas/laminas-mvc": "^3.7",
+        "laminas/laminas-i18n": "^2.0",
+        "laminas/laminas-modulemanager": "^2.16",
+        "laminas/laminas-view": "^2.0",
+        "laminas/laminas-serializer": "^2.0",
+        "laminas/laminas-log": "^2.0",
+        "laminas/laminas-servicemanager": "^3.23",
+        "laminas/laminas-escaper": "^2.0",
+        "casasoft/casamodules": "dev-casawp",
+        "laminas/laminas-mvc-i18n": "^1.9",
+        "laminas/laminas-form": "^3.21"
     },
-    "require-dev": {
-    },
-    "autoload":{
+    "autoload": {
         "psr-4": {
-            "casawp\\": "classes/"
+            "casawp\\": "classes/",
+            "CasasoftStandards\\": "vendor/casasoft/casamodules/CasasoftStandards/src/",
+            "CasasoftMessenger\\": "vendor/casasoft/casamodules/CasasoftMessenger/src/"
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "laminas/laminas-dependency-plugin": true
         }
     }
 }
