@@ -1,6 +1,15 @@
 <?php
 namespace casawp\Service;
 
+use CasasoftStandards\Service\CategoryService;
+use CasasoftStandards\Service\NumvalService;
+use CasasoftMessenger\Service\MessengerService;
+use CasasoftStandards\Service\UtilityService;
+use CasasoftStandards\Service\FeatureService;
+use CasasoftStandards\Service\IntegratedOfferService;
+
+use casawp\Service\FormService;
+
 class OfferService{
     public $post = null;
     private $categories = null; //lazy
@@ -17,15 +26,59 @@ class OfferService{
     private $currentOffer = null;
     private $collection = array();
 
-    public function __construct($categoryService, $numvalService, $messengerService, $utilityService, $featureService, $integratedOfferService, $formService){
-    	$this->utilityService = $utilityService;
-    	$this->categoryService = $categoryService;
-    	$this->numvalService = $numvalService;
-        $this->featureService = $featureService;
-    	$this->messengerService = $messengerService;
-    	$this->integratedOfferService = $integratedOfferService;
-        $this->formService = $formService;
-    }
+	private $utilityService;
+	private $categoryService;
+	private $numvalService;
+	private $featureService;
+	private $messengerService;
+	private $integratedOfferService;
+	private $formService;
+
+	public function __construct(
+		CategoryService $categoryService,
+		NumvalService $numvalService,
+		MessengerService $messengerService,
+		UtilityService $utilityService,
+		FeatureService $featureService,
+		IntegratedOfferService $integratedOfferService,
+		FormService $formService
+	) {
+		$this->utilityService = $utilityService;
+		$this->categoryService = $categoryService;
+		$this->numvalService = $numvalService;
+		$this->featureService = $featureService;
+		$this->messengerService = $messengerService;
+		$this->integratedOfferService = $integratedOfferService;
+		$this->formService = $formService;
+	}
+
+	public function getUtilityService() {
+		return $this->utilityService;
+	}
+
+	public function getCategoryService() {
+		return $this->categoryService;
+	}
+
+	public function getNumvalService() {
+		return $this->numvalService;
+	}
+
+	public function getFeatureService() {
+		return $this->featureService;
+	}
+
+	public function getMessengerService() {
+		return $this->messengerService;
+	}
+
+	public function getIntegratedOfferService() {
+		return $this->integratedOfferService;
+	}
+
+	public function getFormService() {
+		return $this->formService;
+	}
 
 
     //deligate all other gets to the Offer Object 
