@@ -1,21 +1,29 @@
 <?php
 namespace casawp\Service;
 
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
+use Psr\Container\ContainerInterface;
 
 class OfferServiceFactory implements FactoryInterface
 {
-    public function createService(ServiceLocatorInterface $serviceLocator)
+    public function __invoke(ContainerInterface $container, $requestedName, ?array $options = null)
     {
-    	$cs = $serviceLocator->get('CasasoftCategory');
-    	$ns = $serviceLocator->get('CasasoftNumval');
-    	$fs = $serviceLocator->get('CasasoftFeature');
-    	$messenger = $serviceLocator->get('CasasoftMessenger');
-    	$us = $serviceLocator->get('CasasoftUtility');
-        $ios = $serviceLocator->get('CasasoftIntegratedOffer');
-        $formService = $serviceLocator->get('casawpFormService');
-        $service = new OfferService($cs, $ns, $messenger, $us, $fs, $ios, $formService);
-        return $service;
+        $categoryService = $container->get('CasasoftCategory');
+        $numvalService = $container->get('CasasoftNumval');
+        $messengerService = $container->get('CasasoftMessenger');
+        $utilityService = $container->get('CasasoftUtility');
+        $featureService = $container->get('CasasoftFeature');
+        $integratedOfferService = $container->get('CasasoftIntegratedOffer');
+        $formService = $container->get('casawpFormService');
+
+        return new OfferService(
+            $categoryService,
+            $numvalService,
+            $messengerService,
+            $utilityService,
+            $featureService,
+            $integratedOfferService,
+            $formService
+        );
     }
 }
