@@ -1,8 +1,8 @@
 <?php
 namespace casawp\Service;
 
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 
 class ProjectServiceFactory implements FactoryInterface
 {
