@@ -11,12 +11,12 @@
  * EN-Revision: 30.Jul.2011
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Непопълнена стойност на капча",
     "Failed to validate captcha" => "Не може да валидира капча",
     "Captcha value is wrong: %value%" => "Стойността на капча е грешна: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Непопълнена стойност на капча",
     "Captcha ID field is missing" => "Липсващо капча ID поле",
     "Captcha value is wrong" => "Стойността на капча е грешна",
