@@ -11,12 +11,12 @@
  * EN-Revision: 30.Jul.2011
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Falten camps del captcha",
     "Failed to validate captcha" => "No s'ha pogut validar el captcha",
     "Captcha value is wrong: %value%" => "El valor del captcha és incorrecte: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Valor buit del captcha",
     "Captcha ID field is missing" => "Falta el camp ID del captcha",
     "Captcha value is wrong" => "El valor del captcha és incorrecte",
