@@ -11,58 +11,58 @@
  * EN-Revision: 16.Jul.2013
  */
 return array(
-    // Zend\I18n\Validator\Alnum
+    // Laminas\I18n\Validator\Alnum
     "Invalid type given. String, integer or float expected" => "Tipus no vàlid donat. S'espera una cadena de text, un enter o un nombre de precisió simple",
     "The input contains characters which are non alphabetic and no digits" => "L'entrada conté caràcters que no són alfabètics ni dígits",
     "The input is an empty string" => "L'entrada és una cadena buida",
 
-    // Zend\I18n\Validator\Alpha
+    // Laminas\I18n\Validator\Alpha
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
     "The input contains non alphabetic characters" => "L'entrada conté caràcters no alfabètics",
     "The input is an empty string" => "L'entrada és una cadena buida",
 
-    // Zend\I18n\Validator\DateTime
+    // Laminas\I18n\Validator\DateTime
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
     "The input does not appear to be a valid datetime" => "L'entrada no sembla ser un datetime vàlid",
 
-    // Zend\I18n\Validator\Float
+    // Laminas\I18n\Validator\Float
     "Invalid type given. String, integer or float expected" => "Tipus no vàlid donat. S'espera una cadena de text, un enter o un nombre de precisió simple",
     "The input does not appear to be a float" => "L'entrada no sembla ser un nombre de precisió simple",
 
-    // Zend\I18n\Validator\Int
+    // Laminas\I18n\Validator\Int
     "Invalid type given. String or integer expected" => "Tipus no vàlid donat. S'espera una cadena de text o un enter",
     "The input does not appear to be an integer" => "L'entrada no sembla ser un nombre enter",
 
-    // Zend\I18n\Validator\PhoneNumber
+    // Laminas\I18n\Validator\PhoneNumber
     "The input does not match a phone number format" => "L'entrada no coincideix amb un format de número de telèfon",
     "The country provided is currently unsupported" => "El país proporcionat no està suportat",
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
 
-    // Zend\I18n\Validator\PostCode
+    // Laminas\I18n\Validator\PostCode
     "Invalid type given. String or integer expected" => "Tipus no vàlid donat. S'espera una cadena de text o un enter",
     "The input does not appear to be a postal code" => "L'entrada no sembla ser un codi postal",
     "An exception has been raised while validating the input" => "S'ha llançat una excepció en validar l'entrada",
 
-    // Zend\Validator\Barcode
+    // Laminas\Validator\Barcode
     "The input failed checksum validation" => "L'entrada ha fallat la validació de la suma de comprovació",
     "The input contains invalid characters" => "L'entrada conté caràcters no vàlids",
     "The input should have a length of %length% characters" => "L'entrada ha de tenir una longitud de %length% caràcters",
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
 
-    // Zend\Validator\Between
+    // Laminas\Validator\Between
     "The input is not between '%min%' and '%max%', inclusively" => "L'entrada no és entre '%min%' i '%max%', inclusivament",
     "The input is not strictly between '%min%' and '%max%'" => "L'entrada no és estrictament entre '%min%' i '%max%'",
 
-    // Zend\Validator\Bitwise
+    // Laminas\Validator\Bitwise
     "The input has no common bit set with '%control%'" => "L'entrada no té cap bit en comú amb '%control%'",
     "The input doesn't have the same bits set as '%control%'" => "L'entrada no té els mateixos bits que '%control%'",
     "The input has common bit set with '%control%'" => "L'entrada té bits en comú amb '%control%'",
 
-    // Zend\Validator\Callback
+    // Laminas\Validator\Callback
     "The input is not valid" => "L'entrada no és vàlida",
     "An exception has been raised within the callback" => "S'ha llançat una excepció en el callback",
 
-    // Zend\Validator\CreditCard
+    // Laminas\Validator\CreditCard
     "The input seems to contain an invalid checksum" => "L'entrada sembla contenir una suma de comprovació no vàlida",
     "The input must contain only digits" => "L'entrada ha de contenir només dígits",
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
@@ -71,30 +71,30 @@ return array(
     "The input seems to be an invalid credit card number" => "L'entrada sembla ser un número de targeta de crèdit no vàlid",
     "An exception has been raised while validating the input" => "S'ha llançat una excepció validant l'entrada",
 
-    // Zend\Validator\Csrf
+    // Laminas\Validator\Csrf
     "The form submitted did not originate from the expected site" => "El formulari presentat no es va originar en el lloc esperat",
 
-    // Zend\Validator\Date
+    // Laminas\Validator\Date
     "Invalid type given. String, integer, array or DateTime expected" => "Tipus no vàlid donat. S'espera una cadena de text, un enter, un array o DateTime",
     "The input does not appear to be a valid date" => "L'entrada no sembla ser una data vàlida",
     "The input does not fit the date format '%format%'" => "L'entrada no s'ajusta al format de la data '%format%'",
 
-    // Zend\Validator\DateStep
+    // Laminas\Validator\DateStep
     "Invalid type given. String, integer, array or DateTime expected" => "Tipus no vàlid donat. S'espera una cadena de text, un enter, un array o DateTime",
     "The input does not appear to be a valid date" => "L'entrada no sembla ser una data vàlida",
     "The input does not fit the date format '%format%'" => "L'entrada no s'ajusta al format de data '%format%'",
     "The input is not a valid step" => "L'entrada no és un pas vàlid",
 
-    // Zend\Validator\Db\AbstractDb
+    // Laminas\Validator\Db\AbstractDb
     "No record matching the input was found" => "No hi ha cap registre que coincideixi amb l'entrada",
     "A record matching the input was found" => "Es va trobar un registre coincident l'entrada",
 
-    // Zend\Validator\Digits
+    // Laminas\Validator\Digits
     "The input must contain only digits" => "L'entrada només ha de contenir dígits",
     "The input is an empty string" => "L'entrada és una cadena buida",
     "Invalid type given. String, integer or float expected" => "Tipus no vàlid donat. S'espera una cadena de text, un enter o un nombre de precisió simple",
 
-    // Zend\Validator\EmailAddress
+    // Laminas\Validator\EmailAddress
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
     "The input is not a valid email address. Use the basic format local-part@hostname" => "L'entrada no és una adreça vàlida de correu electrònic. Utilitzeu el format bàsic local-part@hostname",
     "'%hostname%' is not a valid hostname for the email address" => "'%hostname%' no és un nom de host vàlid per a la direcció de correu electrònic",
@@ -105,40 +105,40 @@ return array(
     "'%localPart%' is not a valid local part for the email address" => "'%localPart%' no és una part local vàlida per a la direcció de correu electrònic",
     "The input exceeds the allowed length" => "L'entrada supera la longitud permesa",
 
-    // Zend\Validator\Explode
+    // Laminas\Validator\Explode
     "Invalid type given" => "Tipus no vàlid donat",
 
-    // Zend\Validator\File\Count
+    // Laminas\Validator\File\Count
     "Too many files, maximum '%max%' are allowed but '%count%' are given" => "Massa arxius, estan permesos màxim '%max%' però s'han donat '%count%'",
     "Too few files, minimum '%min%' are expected but '%count%' are given" => "Falten arxius, s'espera mínim '%min%' però s'han donat '%count%'",
 
-    // Zend\Validator\File\Crc32
+    // Laminas\Validator\File\Crc32
     "File does not match the given crc32 hashes" => "L'arxiu no coindideix amb el hash crc32 donat",
     "A crc32 hash could not be evaluated for the given file" => "El hash crc32 no es va poder avaluar per l'arxiu donat",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\ExcludeExtension
+    // Laminas\Validator\File\ExcludeExtension
     "File has an incorrect extension" => "L'arxiu té una extensió falsa",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\Exists
+    // Laminas\Validator\File\Exists
     "File does not exist" => "L'arxiu no existeix",
 
-    // Zend\Validator\File\Extension
+    // Laminas\Validator\File\Extension
     "File has an incorrect extension" => "L'arxiu té una extensió falsa",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\FilesSize
+    // Laminas\Validator\File\FilesSize
     "All files in sum should have a maximum size of '%max%' but '%size%' were detected" => "Tots els arxius en la suma haurien de tenir una mida màxima de '%max%' però s'ha detectat la mida '%size%'",
     "All files in sum should have a minimum size of '%min%' but '%size%' were detected" => "Tots els arxius en la suma haurien de tenir una mida màxima de '%max%' però s'ha detectat la mida '%size%'",
     "One or more files can not be read" => "Un o més fitxers no es poden llegir",
 
-    // Zend\Validator\File\Hash
+    // Laminas\Validator\File\Hash
     "File does not match the given hashes" => "L'arxiu no coincideix amb els valors hash donats",
     "A hash could not be evaluated for the given file" => "El hash no es va poder avaluar per l'arxiu donat",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\ImageSize
+    // Laminas\Validator\File\ImageSize
     "Maximum allowed width for image should be '%maxwidth%' but '%width%' detected" => "L'amplada màxima permesa per a la imatge hauria de ser '%maxwidth%' però s'ha detectat '%width%'",
     "Minimum expected width for image should be '%minwidth%' but '%width%' detected" => "L'amplada mínima permesa per a la imatge hauria de ser '%maxwidth%' però s'ha detectat '%width%'",
     "Maximum allowed height for image should be '%maxheight%' but '%height%' detected" => "L'alçada màxima permesa per a la imatge hauria de ser '%maxwidth%' però s'ha detectat '%width%'",
@@ -146,40 +146,40 @@ return array(
     "The size of image could not be detected" => "La mida de la imatge no s'ha pogut detectar",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\IsCompressed
+    // Laminas\Validator\File\IsCompressed
     "File is not compressed, '%type%' detected" => "L'arxiu no està comprimit, s'ha detectat '%type%' ",
     "The mimetype could not be detected from the file" => "El mimetype de l'arxiu no s'ha pogut detectar",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\IsImage
+    // Laminas\Validator\File\IsImage
     "File is no image, '%type%' detected" => "L'arxiu no és una imatge, s'ha detectat '%type%'",
     "The mimetype could not be detected from the file" => "El mimetype de l'arxiu no s'ha pogut detectar",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\Md5
+    // Laminas\Validator\File\Md5
     "File does not match the given md5 hashes" => "L'arxiu no coindideix amb el hash md5 donat",
     "An md5 hash could not be evaluated for the given file" => "El hash md5 no es va poder avaluar per l'arxiu donat",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\MimeType
+    // Laminas\Validator\File\MimeType
     "File has an incorrect mimetype of '%type%'" => "L'arxiu té un mimetype incorrecte del tipus '%type%'",
     "The mimetype could not be detected from the file" => "El mimetype de l'arxiu no s'ha pogut detectar",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\NotExists
+    // Laminas\Validator\File\NotExists
     "File exists" => "L'arxiu existeix",
 
-    // Zend\Validator\File\Sha1
+    // Laminas\Validator\File\Sha1
     "File does not match the given sha1 hashes" => "L'arxiu no coindideix amb el hash sha1 donat",
     "A sha1 hash could not be evaluated for the given file" => "El hash sha1 no es va poder avaluar per l'arxiu donat",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\Size
+    // Laminas\Validator\File\Size
     "Maximum allowed size for file is '%max%' but '%size%' detected" => "La mida màxima permesa per a l'arxiu és '%max%' però s'ha detectat '%size%'",
     "Minimum expected size for file is '%min%' but '%size%' detected" => "La mida mínima permesa per a l'arxiu és '%max%' però s'ha detectat '%size%'",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\File\Upload
+    // Laminas\Validator\File\Upload
     "File '%value%' exceeds the defined ini size" => "L'arxiu '%value%' supera la mida definida inicialment",
     "File '%value%' exceeds the defined form size" => "L'arxiu '%value%' supera la mida definida en el formulari",
     "File '%value%' was only partially uploaded" => "L'arxiu '%value%' s'ha carregat parcialment",
@@ -191,7 +191,7 @@ return array(
     "File '%value%' was not found" => "L'arxiu '%value%' no s'ha trobat",
     "Unknown error while uploading file '%value%'" => "Error desconegut en pujar l'arxiu '%value%'",
 
-    // Zend\Validator\File\UploadFile
+    // Laminas\Validator\File\UploadFile
     "File exceeds the defined ini size" => "L'arxiu supera la mida definida inicialment",
     "File exceeds the defined form size" => "L'arxiu supera la mida definida en el formulari",
     "File was only partially uploaded" => "L'arxiu s'ha carregat parcialment",
@@ -203,20 +203,20 @@ return array(
     "File was not found" => "L'arxiu no s'ha trobat",
     "Unknown error while uploading file" => "Error desconegut en pujar l'arxiu",
 
-    // Zend\Validator\File\WordCount
+    // Laminas\Validator\File\WordCount
     "Too many words, maximum '%max%' are allowed but '%count%' were counted" => "Excés de paraules, màxim '%max%' es permeten però s'han comptat '%count%'",
     "Too few words, minimum '%min%' are expected but '%count%' were counted" => "Falten paraules, mínim '%min%' es permeten però s'han comptat '%count%'",
     "File is not readable or does not exist" => "L'arxiu no és pot llegir o no existeix",
 
-    // Zend\Validator\GreaterThan
+    // Laminas\Validator\GreaterThan
     "The input is not greater than '%min%'" => "L'entrada no és més gran que '%min%'",
     "The input is not greater or equal than '%min%'" => "L'entrada no és més gran o igual que '%min%'",
 
-    // Zend\Validator\Hex
+    // Laminas\Validator\Hex
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
     "The input contains non-hexadecimal characters" => "L'entrada conté caràcters no hexadecimals",
 
-    // Zend\Validator\Hostname
+    // Laminas\Validator\Hostname
     "The input appears to be a DNS hostname but the given punycode notation cannot be decoded" => "L'entrada sembla ser un nom d'amfitrió DNS però la notació punycode donada no pot ser descodificada",
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
     "The input appears to be a DNS hostname but contains a dash in an invalid position" => "L'entrada sembla ser un nom de host DNS, però conté un guió en una posició no vàlida",
@@ -229,69 +229,69 @@ return array(
     "The input appears to be a DNS hostname but cannot extract TLD part" => "L'entrada sembla ser un nom de host DNS però no pot extreure la part TLD",
     "The input appears to be a DNS hostname but cannot match TLD against known list" => "L'entrada sembla ser un nom de host DNS però no s'ha trobat una coincidència del TLD amb la llista coneguda",
 
-    // Zend\Validator\Iban
+    // Laminas\Validator\Iban
     "Unknown country within the IBAN" => "País desconegut dins l'IBAN",
     "Countries outside the Single Euro Payments Area (SEPA) are not supported" => "Els països no pertanyents a la zona única de pagaments (SEPA) no són compatibles",
     "The input has a false IBAN format" => "L'entrada té un fals format IBAN",
     "The input has failed the IBAN check" => "L'entrada no ha passat la verificació IBAN",
 
-    // Zend\Validator\Identical
+    // Laminas\Validator\Identical
     "The two given tokens do not match" => "Els dos tokens donats no coincideixen",
     "No token was provided to match against" => "No s'ha proporcionat cap token per fer la comprovació",
 
-    // Zend\Validator\InArray
+    // Laminas\Validator\InArray
     "The input was not found in the haystack" => "L'entrada no s'ha trobat",
 
-    // Zend\Validator\Ip
+    // Laminas\Validator\Ip
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
     "The input does not appear to be a valid IP address" => "L'entrada no sembla ser una adreça IP vàlida",
 
-    // Zend\Validator\IsInstanceOf
+    // Laminas\Validator\IsInstanceOf
     "The input is not an instance of '%className%'" => "L'entrada no és una instància de '%className%'",
 
-    // Zend\Validator\Isbn
+    // Laminas\Validator\Isbn
     "Invalid type given. String or integer expected" => "Tipus no vàlid donat. S'espera una cadena de text o un enter",
     "The input is not a valid ISBN number" => "L'entrada no és un ISBN vàlid",
 
-    // Zend\Validator\LessThan
+    // Laminas\Validator\LessThan
     "The input is not less than '%max%'" => "L'entrada no és inferior a '%max%'",
     "The input is not less or equal than '%max%'" => "L'entrada no és menor o igual que '%max%'",
 
-    // Zend\Validator\NotEmpty
+    // Laminas\Validator\NotEmpty
     "Value is required and can't be empty" => "El valor és obligatori i no pot estar buit",
     "Invalid type given. String, integer, float, boolean or array expected" => "Tipus no vàlid donat. S'espera una cadena de text, un enter, un nombre de precisió simple, un booleà o un array",
 
-    // Zend\Validator\Regex
+    // Laminas\Validator\Regex
     "Invalid type given. String, integer or float expected" => "Tipus no vàlid donat. S'espera una cadena de text, un enter o un nombre de precisió simple",
     "The input does not match against pattern '%pattern%'" => "L'entrada no coincideix amb el patró '%pattern%'",
     "There was an internal error while using the pattern '%pattern%'" => "S'ha produït un error intern al utilitzar el patró '%pattern%'",
 
-    // Zend\Validator\Sitemap\Changefreq
+    // Laminas\Validator\Sitemap\Changefreq
     "The input is not a valid sitemap changefreq" => "L'entrada no és un mapa de lloc changefreq vàlid",
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
 
-    // Zend\Validator\Sitemap\Lastmod
+    // Laminas\Validator\Sitemap\Lastmod
     "The input is not a valid sitemap lastmod" => "L'entrada no és un mapa de lloc lastmod vàlid",
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
 
-    // Zend\Validator\Sitemap\Loc
+    // Laminas\Validator\Sitemap\Loc
     "The input is not a valid sitemap location" => "L'entrada no és una ubicació del mapa de lloc vàlida",
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
 
-    // Zend\Validator\Sitemap\Priority
+    // Laminas\Validator\Sitemap\Priority
     "The input is not a valid sitemap priority" => "L'entrada no és una prioritat del mapa de lloc vàlida",
     "Invalid type given. Numeric string, integer or float expected" => "Tipus no vàlid donat. S'espera una cadena de text numèrica, un enter o nombre de precisió simple",
 
-    // Zend\Validator\Step
+    // Laminas\Validator\Step
     "Invalid value given. Scalar expected" => "Valor incorrecte donat. S'espera un escalar",
     "The input is not a valid step" => "L'entrada no és un pas vàlid",
 
-    // Zend\Validator\StringLength
+    // Laminas\Validator\StringLength
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
     "The input is less than %min% characters long" => "L'entrada és menor que %min% caràcters",
     "The input is more than %max% characters long" => "L'entrada és més que %max% caràcters",
 
-    // Zend\Validator\Uri
+    // Laminas\Validator\Uri
     "Invalid type given. String expected" => "Tipus no vàlid donat. S'espera una cadena de text",
     "The input does not appear to be a valid Uri" => "L'entrada no sembla ser un URI vàlid",
 );
