@@ -11,12 +11,12 @@
  * CS-Revision: 4.Nov.2013
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Chybí pole captcha",
     "Failed to validate captcha" => "Captchu se nepodařilo ověřit",
     "Captcha value is wrong: %value%" => "Hodnota captchy není správná: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Hodnota captcha nebyla zadána",
     "Captcha ID field is missing" => "Chybí pole captcha ID",
     "Captcha value is wrong" => "Hodnota captchy není správná",
