@@ -11,12 +11,12 @@
  * EN-Revision: 30.Jul.2011
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Missing captcha fields",
     "Failed to validate captcha" => "Failed to validate captcha",
     "Captcha value is wrong: %value%" => "Captcha value is wrong: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Empty captcha value",
     "Captcha ID field is missing" => "Captcha ID field is missing",
     "Captcha value is wrong" => "Captcha value is wrong",
