@@ -11,12 +11,12 @@
  * EN-Revision: 30.Jul.2011
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Faltan campos del captcha",
     "Failed to validate captcha" => "La validación del captcha falló",
     "Captcha value is wrong: %value%" => "El valor del captcha es erróneo: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Valor del captcha vacío",
     "Captcha ID field is missing" => "Falta ell campo ID de captcha",
     "Captcha value is wrong" => "El valor del captcha es erróneo",
