@@ -11,12 +11,12 @@
  * ID-Revision: 08.Dec.2013
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Tidak ada bidang captcha",
     "Failed to validate captcha" => "Gagal memvalidasi captcha",
     "Captcha value is wrong: %value%" => "Nilai captcha salah: %value%",
 
-    // Zend\Captcha\AbstractWord
+    // Laminas\Captcha\AbstractWord
     "Empty captcha value" => "Nilai captcha kosong",
     "Captcha ID field is missing" => "Bidang Captcha ID tidak ada",
     "Captcha value is wrong" => "Nilai Captcha salah",
