@@ -11,71 +11,71 @@
  * ID-Revision: 08.Dec.2013
  */
 return array(
-    // Zend\I18n\Validator\Alnum
+    // Laminas\I18n\Validator\Alnum
     "Invalid type given. String, integer or float expected" => "Tipe yang diberikan tidak valid. Diharapkan String, integer, atau float",
     "The input contains characters which are non alphabetic and no digits" => "Isian berisi karakter yang non abjad dan tidak ada angka",
     "The input is an empty string" => "Isian adalah string kosong",
 
-    // Zend\I18n\Validator\Alpha
+    // Laminas\I18n\Validator\Alpha
     "Invalid type given. String expected" => "Tipe yang diberikan tidak valid. Diharapkan string",
     "The input contains non alphabetic characters" => "Isian berisi karakter non abjad",
 
-    // Zend\I18n\Validator\DateTime
+    // Laminas\I18n\Validator\DateTime
     "The input does not appear to be a valid datetime" => "Isian tidak tampak sebagai datetime yang valid",
 
-    // Zend\I18n\Validator\IsFloat
+    // Laminas\I18n\Validator\IsFloat
     "The input does not appear to be a float" => "Isian tidak tampak sebagai float",
 
-    // Zend\I18n\Validator\IsInt
+    // Laminas\I18n\Validator\IsInt
     "Invalid type given. String or integer expected" => "Tipe yang diberikan tidak valid. Diharapkan string atau integer",
     "The input does not appear to be an integer" => "Isian tidak tampak sebagai integer",
 
-    // Zend\I18n\Validator\PhoneNumber
+    // Laminas\I18n\Validator\PhoneNumber
     "The input does not match a phone number format" => "Isian tidak cocok dengan format nomor telepon",
     "The country provided is currently unsupported" => "Negara yang diisikan saat ini tidak didukung",
 
-    // Zend\I18n\Validator\PostCode
+    // Laminas\I18n\Validator\PostCode
     "The input does not appear to be a postal code" => "Isian tidak tampak sebagai kode pos",
     "An exception has been raised while validating the input" => "Eksepsi muncul pada saat memvalidasi isian",
 
-    // Zend\Validator\Barcode
+    // Laminas\Validator\Barcode
     "The input failed checksum validation" => "Isian gagal memvalidasi checksum",
     "The input contains invalid characters" => "Isian mengandung karakter yang tidak valid",
     "The input should have a length of %length% characters" => "Isian harus memiliki panjang panjang %length% karakter",
 
-    // Zend\Validator\Between
+    // Laminas\Validator\Between
     "The input is not between '%min%' and '%max%', inclusively" => "Isian tidak berada di antara '%min%' and '%max%', inklusif",
     "The input is not strictly between '%min%' and '%max%'" => "Isian tidak ketat berada di antara '%min%' dan '%max%'",
 
-    // Zend\Validator\Callback
+    // Laminas\Validator\Callback
     "The input is not valid" => "Isian tidak valid",
     "An exception has been raised within the callback" => "Eksepsi muncul dalam callback",
 
-    // Zend\Validator\CreditCard
+    // Laminas\Validator\CreditCard
     "The input seems to contain an invalid checksum" => "Isian tampaknya mengandung checksum yang tidak valid",
     "The input must contain only digits" => "Isian harus berisi angka saja",
     "The input contains an invalid amount of digits" => "Isian berisi jumlah digit yang tidak valid",
     "The input is not from an allowed institute" => "Isian tidak berasal dari lembaga yang diperbolehkan",
     "The input seems to be an invalid credit card number" => "Isian tampaknya nomor kartu kredit yang tidak valid",
 
-    // Zend\Validator\Csrf
+    // Laminas\Validator\Csrf
     "The form submitted did not originate from the expected site" => "Form yang disubmit tidak berasal dari situs yang diharapkan",
 
-    // Zend\Validator\Date
+    // Laminas\Validator\Date
     "Invalid type given. String, integer, array or DateTime expected" => "Tipe yang diberikan tidak valid. Diharapkan String, integer, array, atau DateTime",
     "The input does not appear to be a valid date" => "Isian tidak tampak sebagai tanggal yang valid",
     "The input does not fit the date format '%format%'" => "Isian tidak sesuai dengan format tangal '%format%'",
 
-    // Zend\Validator\DateStep
+    // Laminas\Validator\DateStep
     "The input is not a valid step" => "Isian bukan langkah yang valid",
 
-    // Zend\Validator\Db\AbstractDb
+    // Laminas\Validator\Db\AbstractDb
     "No record matching the input was found" => "Tidak ada catatan pencocokan isian yang ditemukan",
     "A record matching the input was found" => "Sebuah catatan pencocokan isian ditemukan",
 
-    // Zend\Validator\Digits
+    // Laminas\Validator\Digits
 
-    // Zend\Validator\EmailAddress
+    // Laminas\Validator\EmailAddress
     "The input is not a valid email address. Use the basic format local-part@hostname" => "Isian bukan alamat email yang valid. Gunakan format bagian-lokal@hostname",
     "'%hostname%' is not a valid hostname for the email address" => "'%hostname%' adalah hostname yang tidak valid untuk alamat email",
     "'%hostname%' does not appear to have any valid MX or A records for the email address" => "'%hostname%' tidak tampak untuk memiliki MX atau A records yang valid untuk alamat email",
@@ -85,67 +85,67 @@ return array(
     "'%localPart%' is not a valid local part for the email address" => "'%localPart%' adalah bagian local yang tidak valid untuk email address",
     "The input exceeds the allowed length" => "Isian melebihi panjang yang diperbolehkan",
 
-    // Zend\Validator\Explode
+    // Laminas\Validator\Explode
     "Invalid type given" => "Tipe yang diberikan tidak valid",
 
-    // Zend\Validator\File\Count
+    // Laminas\Validator\File\Count
     "Too many files, maximum '%max%' are allowed but '%count%' are given" => "Terlalu banyak berkas, maksimal '%max%' diperbolehkan tetapi '%count%' diberikan",
     "Too few files, minimum '%min%' are expected but '%count%' are given" => "Terlalu sedikit berkas, minimal '%min%' diperbolehkan tetapi '%count%' diberikan",
 
-    // Zend\Validator\File\Crc32
+    // Laminas\Validator\File\Crc32
     "File does not match the given crc32 hashes" => "Berkas tidak cocok dengan hash crc32 diberikan",
     "A crc32 hash could not be evaluated for the given file" => "Sebuah hash crc32 tidak dapat dievaluasi untuk berkas yang diberikan",
     "File is not readable or does not exist" => "Berkas tidak terbaca atau tidak ada",
 
-    // Zend\Validator\File\ExcludeExtension
+    // Laminas\Validator\File\ExcludeExtension
     "File has an incorrect extension" => "Ekstensi berkas salah",
 
-    // Zend\Validator\File\Exists
+    // Laminas\Validator\File\Exists
     "File does not exist" => "Berkas tidak ada",
 
-    // Zend\Validator\File\Extension
+    // Laminas\Validator\File\Extension
 
-    // Zend\Validator\File\FilesSize
+    // Laminas\Validator\File\FilesSize
     "All files in sum should have a maximum size of '%max%' but '%size%' were detected" => "Semua berkas yang dijumlahkan harus memiliki ukuran maksimal '%max%' tapi ukuran '%size%' terdeteksi",
     "All files in sum should have a minimum size of '%min%' but '%size%' were detected" => "Semua berkas yang dijumlahkan harus memiliki ukuran minimal '%min%' tapi ukuran '%size%' terdeteksi",
     "One or more files can not be read" => "Satu atau beberapa berkas tidak dapat dibaca",
 
-    // Zend\Validator\File\Hash
+    // Laminas\Validator\File\Hash
     "File does not match the given hashes" => "Berkas tidak cocok dengan hash yang diberikan",
     "A hash could not be evaluated for the given file" => "Sebuah hash tidak dapat dievaluasi untuk berkas yang diberikan",
 
-    // Zend\Validator\File\ImageSize
+    // Laminas\Validator\File\ImageSize
     "Maximum allowed width for image should be '%maxwidth%' but '%width%' detected" => "Lebar maksimal yang diharapkan untuk gambar seharusnya '%maxwidth%' tetapi '%width%' terdeteksi",
     "Minimum expected width for image should be '%minwidth%' but '%width%' detected" => "Lebar minimal yang diharapkan untuk gambar seharusnya '%minwidth%' tetapi '%width%' terdeteksi",
     "Maximum allowed height for image should be '%maxheight%' but '%height%' detected" => "Tinggi maksimal yang diharapkan untuk gambar seharusnya '%maxheight%' tetapi '%height%' terdeteksi",
     "Minimum expected height for image should be '%minheight%' but '%height%' detected" => "Tinggi minimal yang diharapkan untuk gambar seharusnya '%minheight%' tetapi '%height%' terdeteksi",
     "The size of image could not be detected" => "Ukuran gambar tidak dapat terdeteksi",
 
-    // Zend\Validator\File\IsCompressed
+    // Laminas\Validator\File\IsCompressed
     "File is not compressed, '%type%' detected" => "Bukan berkas terkompresi, '%type%' terdeteksi",
     "The mimetype could not be detected from the file" => "Mimetype tidak dapat terdeteksi dari berkas",
 
-    // Zend\Validator\File\IsImage
+    // Laminas\Validator\File\IsImage
     "File is no image, '%type%' detected" => "Berkas bukan gambar, '%type%' terdeteksi",
 
-    // Zend\Validator\File\Md5
+    // Laminas\Validator\File\Md5
     "File does not match the given md5 hashes" => "Berkas tidak cocok dengan hash md5 diberikan",
     "An md5 hash could not be evaluated for the given file" => "Md5 hash tidak dapat dievaluasi dari berkas yang diberikan",
 
-    // Zend\Validator\File\MimeType
+    // Laminas\Validator\File\MimeType
     "File has an incorrect mimetype of '%type%'" => "Berkas mempunyai mimetype yang tidak cocok dengan '%type%'",
 
-    // Zend\Validator\File\NotExists
+    // Laminas\Validator\File\NotExists
 
-    // Zend\Validator\File\Sha1
+    // Laminas\Validator\File\Sha1
     "File does not match the given sha1 hashes" => "Berkas tidak cocok dengan hash sha1 diberikan",
     "A sha1 hash could not be evaluated for the given file" => "Sebuah hash sha1 tidak dapat dievaluasi untuk berkas yang diberikan",
 
-    // Zend\Validator\File\Size
+    // Laminas\Validator\File\Size
     "Maximum allowed size for file is '%max%' but '%size%' detected" => "Ukuran maksimal yang diharapkan adalah '%min%' tetapi '%size%' terdeteksi",
     "Minimum expected size for file is '%min%' but '%size%' detected" => "Ukuran minimal yang diharapkan adalah '%min%' tetapi '%size%' terdeteksi",
 
-    // Zend\Validator\File\Upload
+    // Laminas\Validator\File\Upload
     "File '%value%' exceeds the defined ini size" => "Berkas '%value%' melebihi ukuran ini yang didefinisikan",
     "File '%value%' exceeds the defined form size" => "Berkas '%value%' melebihi ukuran form yang didefinisikan",
     "File '%value%' was only partially uploaded" => "Berkas '%value%' hanya sebagian terupload",
@@ -157,7 +157,7 @@ return array(
     "File '%value%' was not found" => "Berkas '%value%' tidak ditemukan",
     "Unknown error while uploading file '%value%'" => "Galat tidak diketahui pada saat proses mengupload berkas '%value%'",
 
-    // Zend\Validator\File\UploadFile
+    // Laminas\Validator\File\UploadFile
     "File exceeds the defined ini size" => "Berkas melebihi ukuran ini yang didefinisikan",
     "File exceeds the defined form size" => "Berkas melebihi ukuran form yang didefinisikan",
     "File was only partially uploaded" => "Berkas hanya sebagian terupload",
@@ -169,18 +169,18 @@ return array(
     "File was not found" => "Berkas tidak ditemukan",
     "Unknown error while uploading file" => "Galat tidak diketahui pada saat unggah berkas",
 
-    // Zend\Validator\File\WordCount
+    // Laminas\Validator\File\WordCount
     "Too many words, maximum '%max%' are allowed but '%count%' were counted" => "Terlalu banyak kata, maksimal '%min%' diharapkan tetapi '%count%' terdeteksi",
     "Too few words, minimum '%min%' are expected but '%count%' were counted" => "Terlalu sedikit kata, minimal '%min%' diharapkan tetapi '%count%' terdeteksi",
 
-    // Zend\Validator\GreaterThan
+    // Laminas\Validator\GreaterThan
     "The input is not greater than '%min%'" => "Isian tidak lebih besar dari '%min%'",
     "The input is not greater or equal than '%min%'" => "Isian tidak lebih besar atau sama dengan '%min%'",
 
-    // Zend\Validator\Hex
+    // Laminas\Validator\Hex
     "The input contains non-hexadecimal characters" => "Isian berisi karakter non-heksadesimal",
 
-    // Zend\Validator\Hostname
+    // Laminas\Validator\Hostname
     "The input appears to be a DNS hostname but the given punycode notation cannot be decoded" => "Isian tampaknya menjadi nama host DNS tetapi notasi punycode diberikan tidak dapat diterjemahkan",
     "The input appears to be a DNS hostname but contains a dash in an invalid position" => "Isian tampaknya menjadi nama host DNS tetapi berisi dash dalam posisi yang tidak valid",
     "The input does not match the expected structure for a DNS hostname" => "Isian tidak cocok dengan struktur untuk hostname DNS yang diharapkan",
@@ -192,65 +192,65 @@ return array(
     "The input appears to be a DNS hostname but cannot extract TLD part" => "Isian tampaknya menjadi hostname DNS tetapi tidak dapat mengekstrak bagian TLD",
     "The input appears to be a DNS hostname but cannot match TLD against known list" => "Isian tampaknya menjadi hostname DNS tapi tidak bisa cocok dengan TLD yang dikenal",
 
-    // Zend\Validator\Iban
+    // Laminas\Validator\Iban
     "Unknown country within the IBAN" => "Negara tidak diketahui dalam IBAN",
     "Countries outside the Single Euro Payments Area (SEPA) are not supported" => "Negara-negara di luar Single Euro Payments Area (SEPA) tidak didukung",
     "The input has a false IBAN format" => "Isian mempunyai IBAN format yang salah",
     "The input has failed the IBAN check" => "Isian telah gagal cek IBAN",
 
-    // Zend\Validator\Identical
+    // Laminas\Validator\Identical
     "The two given tokens do not match" => "Dua token yang diberikan tidak cocok",
     "No token was provided to match against" => "Tidak ada token yang diberikan untuk dibandingkan",
 
-    // Zend\Validator\InArray
+    // Laminas\Validator\InArray
     "The input was not found in the haystack" => "Isian tidak ditemukan di dalam haystack",
 
-    // Zend\Validator\Ip
+    // Laminas\Validator\Ip
     "The input does not appear to be a valid IP address" => "Isian tidak tampak sebagai IP address yang valid",
 
-    // Zend\Validator\IsInstanceOf
+    // Laminas\Validator\IsInstanceOf
     "The input is not an instance of '%className%'" => "Isian bukanlah instance dari '%ClassName%'",
 
-    // Zend\Validator\Isbn
+    // Laminas\Validator\Isbn
     "The input is not a valid ISBN number" => "Isian adalah ISBN number yang tidak valid",
 
-    // Zend\Validator\LessThan
+    // Laminas\Validator\LessThan
     "The input is not less than '%max%'" => "Isian tidak kurang dari '%max%'",
     "The input is not less or equal than '%max%'" => "Isian tidak kurang atau sama dengan '%max%'",
 
-    // Zend\Validator\NotEmpty
+    // Laminas\Validator\NotEmpty
     "Value is required and can't be empty" => "Nilai diperlukan dan tidak boleh kosong",
     "Invalid type given. String, integer, float, boolean or array expected" => "Tipe tidak valid. Diharapkan string, integer, float, boolean, atau array",
 
-    // Zend\Validator\Regex
+    // Laminas\Validator\Regex
     "The input does not match against pattern '%pattern%'" => "Isian tidak cocok dengan pola '%pattern%'",
     "There was an internal error while using the pattern '%pattern%'" => "Terjadi galat internal saat menggunakan pola '%pattern%'",
 
-    // Zend\Validator\Sitemap\Changefreq
+    // Laminas\Validator\Sitemap\Changefreq
     "The input is not a valid sitemap changefreq" => "Isian adalah changefreq sitemap yang tidak valid",
 
-    // Zend\Validator\Sitemap\Lastmod
+    // Laminas\Validator\Sitemap\Lastmod
     "The input is not a valid sitemap lastmod" => "Isian adalah lastmod sitemap yang tidak valid",
 
-    // Zend\Validator\Sitemap\Loc
+    // Laminas\Validator\Sitemap\Loc
     "The input is not a valid sitemap location" => "Isian adalah lokasi sitemap yang tidak valid",
 
-    // Zend\Validator\Sitemap\Priority
+    // Laminas\Validator\Sitemap\Priority
     "The input is not a valid sitemap priority" => "Isian adalah prioritas sitemap yang tidak valid",
     "Invalid type given. Numeric string, integer or float expected" => "Tipe yang diberikan tidak valid. Diharapkan numeric string, integer, atau float",
 
-    // Zend\Validator\Step
+    // Laminas\Validator\Step
     "Invalid value given. Scalar expected" => "Tipe yang diberikan tidak valid. Diharapkan scalar",
 
-    // Zend\Validator\StringLength
+    // Laminas\Validator\StringLength
     "The input is less than %min% characters long" => "Isian kurang dari %min% panjang karakter",
     "The input is more than %max% characters long" => "Isian lebih dari %max% panjang karakter",
 
-    // Zend\Validator\Timezone
+    // Laminas\Validator\Timezone
     "Invalid timezone given." => "Zona waktu yang diberikan tidak valid.",
     "Invalid timezone location given." => "Lokasi zona waktu yang diberikan tidak valid.",
     "Invalid timezone abbreviation given." => "Singkatan zona waktu yang diberikan tidak valid.",
 
-    // Zend\Validator\Uri
+    // Laminas\Validator\Uri
     "The input does not appear to be a valid Uri" => "Isian tidak tampak sebagai Uri yang valid",
 );
