@@ -11,12 +11,12 @@
  * EN-Revision: 30.Jul.2011
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Campi captcha mancanti",
     "Failed to validate captcha" => "Validazione del captcha fallito",
     "Captcha value is wrong: %value%" => "Il valore del Captcha è sbagliato: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Valore del captcha vuoto",
     "Captcha ID field is missing" => "Manca il campo Captcha ID",
     "Captcha value is wrong" => "Il valore del Captcha è sbagliato",
