@@ -11,12 +11,12 @@
  * EN-Revision: 30.Jul.2011
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "キャプチャフィールドが足りません",
     "Failed to validate captcha" => "キャプチャの検証に失敗しました",
     "Captcha value is wrong: %value%" => "キャプチャの値が間違っています: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "キャプチャの値が空です",
     "Captcha ID field is missing" => "キャプチャIDフィールドが足りません",
     "Captcha value is wrong" => "キャプチャの値が間違っています",
