@@ -11,12 +11,12 @@
  * EN-Revision: 16.May.2014
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Не хватает полей captcha",
     "Failed to validate captcha" => "Captcha не прошла проверку",
     "Captcha value is wrong: %value%" => "Не корректное значение captcha: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Пустое значение captcha",
     "Captcha ID field is missing" => "Не хватает поля captcha ID",
     "Captcha value is wrong" => "Не корректное значение captcha",
