@@ -11,12 +11,12 @@
  * SK-Revision: 30.Sep.2013
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Chýba captcha pole",
     "Failed to validate captcha" => "Captchu sa nepodarilo overiť",
     "Captcha value is wrong: %value%" => "Hodnota captchy neni správna: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Hodnota captchy nebola zadaná",
     "Captcha ID field is missing" => "Chýba pole captcha ID",
     "Captcha value is wrong" => "Chybná hodnota catpcha",
