@@ -11,12 +11,12 @@
  * EN-Revision: 30.Jul.2011
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Manjka varnostna koda",
     "Failed to validate captcha" => "Varnostne kode ni bilo mogoče preveriti",
     "Captcha value is wrong: %value%" => "Napačna varnostna koda: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Prazna varnostna koda",
     "Captcha ID field is missing" => "Polje ID varnostne kode manjka",
     "Captcha value is wrong" => "Varnostna koda je napačna",
