@@ -11,12 +11,12 @@
  * EN-Revision: 30.Jul.2011
  */
 return array(
-    // Zend\Captcha\ReCaptcha
+    // Laminas\Captcha\ReCaptcha
     "Missing captcha fields" => "Бракує полів captcha",
     "Failed to validate captcha" => "Captcha не пройшла перевірку",
     "Captcha value is wrong: %value%" => "Неправильне значення captcha: %value%",
 
-    // Zend\Captcha\Word
+    // Laminas\Captcha\Word
     "Empty captcha value" => "Порожнє значення captcha",
     "Captcha ID field is missing" => "Бракує поля captcha ID",
     "Captcha value is wrong" => "Неправильне значення captcha",
